Adds Tailwind and a recursive vertical stack.
I also added a view all link to the project section, which is where I was focusing.

I added tailwind because I decided I didn't want to make up all my utility classes, but to use the prefix utw- to combine both a u and tw prefix.

As a side effect, it appears that though I haven't explicitly loaded the Tailwind reset, it is being applied and changed all my fonts. I actually rather like the links better this way, but the lighter weight headers should be reversed.